mcp_servers/hunter/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
mcp>=0.1.0
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0 